+++ src/actions/BackupActions.js
@@ -5,10 +5,11 @@ import { _loadDataFromFile, _saveDataToFile } from 'actions/AppActions';
 import { updateSettings } from 'actions/SettingActions';
 import { updateProcess } from 'actions/ThreadActions';
 import { getSettings } from 'selectors/SettingSelectors';
-import { exists, join } from 'utils/ElectronUtils';
+import { exists, joinSync } from 'utils/ElectronIpc';
 
 export async function getBackupIds(settings) {
-    const path = join(getDataFolder(settings), 'backups');
+    const dataFolder = await getDataFolder(settings);
+    const path = joinSync(dataFolder, 'backups');
 
     let items = await readDirectory(path);
     items = items.map(item => item.replace('.zip', ''));
@@ -19,16 +20,17 @@ export async function getBackupIds(settings) {
 
 export function restoreBackup(backupId) {
     return async (dispatch, getState) => {
-        let path = join(getDataFolder(getState().settings), 'backups');
+        const dataFolder = await getDataFolder(getState().settings);
+        let path = joinSync(dataFolder, 'backups');
 
         let zip;
 
         try {
-            await exists(join(path, backupId + '.zip'));
-            path = join(path, backupId + '.zip');
+            await exists(joinSync(path, backupId + '.zip'));
+            path = joinSync(path, backupId + '.zip');
             zip = true;
         } catch (e) {
-            path = join(path, backupId);
+            path = joinSync(path, backupId);
             zip = false;
         }
 
@@ -38,7 +40,8 @@ export function restoreBackup(backupId) {
 
 export function backupData() {
     return async (dispatch, getState) => {
-        const path = join(getDataFolder(getState().settings), 'backups', moment.utc().format('YYYYMMDD[T]HHmmss[Z]') + '.zip');
+        const dataFolder = await getDataFolder(getState().settings);
+        const path = joinSync(dataFolder, 'backups', moment.utc().format('YYYYMMDD[T]HHmmss[Z]') + '.zip');
         await dispatch(_saveDataToFile(path, { clean: false, message: 'Backup database', zip: true }));
         await dispatch(cleanBackups());
 
@@ -60,14 +63,14 @@ export function deleteBackup(backupId) {
         }));
 
         try {
-            const dataFolder = getDataFolder(getSettings(state));
-            const path = join(dataFolder, 'backups');
+            const dataFolder = await getDataFolder(getSettings(state));
+            const path = joinSync(dataFolder, 'backups');
 
             try {
-                await exists(join(path, backupId + '.zip'));
-                await deletePath(join(path, backupId + '.zip'), dataFolder);
+                await exists(joinSync(path, backupId + '.zip'));
+                await deletePath(joinSync(path, backupId + '.zip'), dataFolder);
             } catch (e) {
-                await deletePath(join(path, backupId), dataFolder);
+                await deletePath(joinSync(path, backupId), dataFolder);
             }
 
             dispatch(updateProcess({