+++ src/components/notes/common/NoteTitle.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Icon from 'components/common/Icon';
 import { useNote } from 'hooks/UseNote';
 
-export function NoteTitle(props) {
+function NoteTitle(props) {
     const note = useNote(props.noteId);
     return note ? <Icon icon="circle" color={note.color} text={note.title} /> : <span>&nbsp;</span>;
 }