+++ src/components/sorters/SortDirectionTitle.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import Icon from 'components/common/Icon';
 import { useSortDirectionApi } from 'hooks/UseSortDirectionApi';
 
-export function SortDirectionTitle(props) {
+function SortDirectionTitle(props) {
     const sortDirectionApi = useSortDirectionApi();
     const sortDirection = sortDirectionApi.sortDirections.find(sortDirection => sortDirection.id === props.sortDirectionId);