+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## New Features
 
-* Quick sort by clicking on a column header
+* Upgrade to Electron 10
+* Quick sort by clicking on a column header (#185)
+* Set priority and status in "task quick add" bar (#216)
 
 ## Bug Fixes